comment delete fix

+++ board-api/src/main/java/com/douzone/board/service/comments/CommentServiceImpl.java
@@ -1,5 +1,6 @@
 package com.douzone.board.service.comments;
 
+import com.auth0.jwt.JWT;
 import com.douzone.board.entity.Comment;
 import com.douzone.board.repository.BoardRepository;
 import com.douzone.board.repository.CommentRepository;
@@ -7,11 +8,20 @@ import com.douzone.board.repository.CommentRepository;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import com.douzone.board.repository.UserRepository;
 import com.douzone.board.web.dto.CommentDto;
 import lombok.RequiredArgsConstructor;
+import org.apache.tomcat.util.http.parser.Authorization;
+import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 @Service
 @RequiredArgsConstructor
@@ -59,7 +69,17 @@ public class CommentServiceImpl implements CommentService {
 	}
 
 	@Override
-	public void removeComment(Long commentId) {
-		commentRepository.deleteById(commentId);
+	public void removeComment(Long commentId, HttpServletRequest request, HttpServletResponse response) {
+		String token = request.getHeader(AUTHORIZATION);
+		String encodedToken = token.split("Bearer ")[1];
+		String username = JWT.decode(encodedToken).getSubject();
+
+		Comment comment = commentRepository.findById(commentId).orElseThrow(IllegalAccessError::new);
+
+		if(!Objects.equals(username, comment.getUser().getUsername())){
+			response.setStatus(400);
+		}else {
+			commentRepository.deleteById(commentId);
+		}
 	}
 }
